+++ packages/stacked_tools/bin/stacked.dart
@@ -1,17 +1,17 @@
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
-import 'package:stackedcli/src/commands/compile/compile_command.dart';
-import 'package:stackedcli/src/commands/create/create_command.dart';
-import 'package:stackedcli/src/exceptions/invalid_stacked_structure_exception.dart';
-import 'package:stackedcli/src/locator.dart';
+import 'package:stacked_tools/src/commands/compile/compile_command.dart';
+import 'package:stacked_tools/src/commands/create/create_command.dart';
+import 'package:stacked_tools/src/exceptions/invalid_stacked_structure_exception.dart';
+import 'package:stacked_tools/src/locator.dart';
 
 Future<void> main(List<String> arguments) async {
   await setupLocator();
 
   try {
     CommandRunner(
-      'stackedcli',
+      'stacked_tools',
       'A command line interface for building and scaffolding stacked apps',
     )
       ..addCommand(CreateCommand())