+++ packages/stacked_tools/lib/src/commands/create/create_app_command.dart
@@ -3,9 +3,9 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
-import 'package:stackedcli/src/locator.dart';
-import 'package:stackedcli/src/services/template_service.dart';
-import 'package:stackedcli/src/templates/template_constants.dart';
+import 'package:stacked_tools/src/locator.dart';
+import 'package:stacked_tools/src/services/template_service.dart';
+import 'package:stacked_tools/src/templates/template_constants.dart';
 
 class CreateAppCommand extends Command {
   final _templateService = locator<TemplateService>();