+++ packages/stacked_tools/lib/src/commands/create/create_command.dart
@@ -1,7 +1,7 @@
 import 'package:args/command_runner.dart';
-import 'package:stackedcli/src/commands/create/create_app_command.dart';
-import 'package:stackedcli/src/commands/create/create_service_command.dart';
-import 'package:stackedcli/src/commands/create/create_view_command.dart';
+import 'package:stacked_tools/src/commands/create/create_app_command.dart';
+import 'package:stacked_tools/src/commands/create/create_service_command.dart';
+import 'package:stacked_tools/src/commands/create/create_view_command.dart';
 
 /// A command with subcommands that allows you to create / scaffold
 /// different parts of the stacked application