+++ packages/stacked_tools/lib/src/locator.dart
@@ -1,9 +1,9 @@
 import 'package:get_it/get_it.dart';
-import 'package:stackedcli/src/services/file_service.dart';
-import 'package:stackedcli/src/services/path_service.dart';
-import 'package:stackedcli/src/services/pubspec_service.dart';
-import 'package:stackedcli/src/services/template_service.dart';
-import 'package:stackedcli/src/templates/template_helper.dart';
+import 'package:stacked_tools/src/services/file_service.dart';
+import 'package:stacked_tools/src/services/path_service.dart';
+import 'package:stacked_tools/src/services/pubspec_service.dart';
+import 'package:stacked_tools/src/services/template_service.dart';
+import 'package:stacked_tools/src/templates/template_helper.dart';
 
 final locator = GetIt.I;