+++ packages/stacked_tools/lib/src/templates/template_constants.dart
@@ -68,10 +68,10 @@ const String k{{name}}Template{{fileName}}Content = \'''
 
 // TODO: add an option to create Cupertino or custom routes
 const String kTemplateMapDataStructure = '''
-import 'package:stackedcli/src/message_constants.dart';
-import 'package:stackedcli/src/models/template_models.dart';
-import 'package:stackedcli/src/templates/compiled_templates.dart';
-import 'package:stackedcli/src/templates/template_constants.dart';
+import 'package:stacked_tools/src/message_constants.dart';
+import 'package:stacked_tools/src/models/template_models.dart';
+import 'package:stacked_tools/src/templates/compiled_templates.dart';
+import 'package:stacked_tools/src/templates/template_constants.dart';
 
 Map<String, StackedTemplate> kCompiledStackedTemplates = {
   {{#stackedTemplates}}