+++ packages/stacked_tools/CHANGELOG.md
@@ -3,6 +3,6 @@
 ### Initial Release
 
 This version of the cli has 3 create commands:
-- stackedcli create app: This creates a brand new flutter project with all the stacked functionality setup
-- stackedcli create view: This creates a new view in the project and makes all required changes
-- stackedcli create service: This creates a new service in the project and makes all required changes
+- stacked create app: This creates a brand new flutter project with all the stacked functionality setup
+- stacked create view: This creates a new view in the project and makes all required changes
+- stacked create service: This creates a new service in the project and makes all required changes